workflow/requirements.txt
Flask=
psycopg2-binary
langchain
openai
requests
python-dotenv